feat(payroll): implement last accessed month storage for Payroll and Salary Report pages
