+++ src/pages/Payroll/PayrollPage.tsx
@@ -47,11 +47,13 @@ import MonthNavigator from "../../components/MonthNavigator";
 import PayrollUnifiedTable from "../../components/Payroll/PayrollUnifiedTable";
 import PayrollSectionPrintMenu from "../../components/Payroll/PayrollSectionPrintMenu";
 import { useScrollRestoration } from "../../hooks/useScrollRestoration";
+import {
+  readLastAccessedPayrollMonth,
+  saveLastAccessedPayrollMonth,
+} from "../../utils/payroll/payrollMonthStorage";
 
 const FIRST_WEEK_DAY_OF_MONTH: number = 7;
 const EXPANDED_JOBS_STORAGE_PREFIX: string = "payroll-expanded-jobs:";
-const LAST_ACCESSED_PAYROLL_MONTH_STORAGE_KEY: string =
-  "payroll-last-accessed-month";
 const SEARCH_TERM_STORAGE_KEY: string = "payroll-search-term";
 const CLEAR_SEARCH_ON_RETURN_STORAGE_KEY: string =
   "payroll-clear-search-on-return";
@@ -62,39 +64,6 @@ const getDefaultPayrollMonth = (today: Date = new Date()): Date => {
     today.getDate() <= FIRST_WEEK_DAY_OF_MONTH ? -1 : 0;
 
   return new Date(today.getFullYear(), today.getMonth() + monthOffset, 1);
-};
-
-const readLastAccessedPayrollMonth = (): Date | null => {
-  try {
-    const storedMonth: string | null = localStorage.getItem(
-      LAST_ACCESSED_PAYROLL_MONTH_STORAGE_KEY
-    );
-    const matchedMonth: RegExpExecArray | null = storedMonth
-      ? /^(\d{4})-(0[1-9]|1[0-2])$/.exec(storedMonth)
-      : null;
-
-    if (!matchedMonth) return null;
-
-    const year: number = Number(matchedMonth[1]);
-    const month: number = Number(matchedMonth[2]);
-
-    return new Date(year, month - 1, 1);
-  } catch {
-    return null;
-  }
-};
-
-const saveLastAccessedPayrollMonth = (selectedMonth: Date): void => {
-  try {
-    const year: number = selectedMonth.getFullYear();
-    const month: string = String(selectedMonth.getMonth() + 1).padStart(2, "0");
-    localStorage.setItem(
-      LAST_ACCESSED_PAYROLL_MONTH_STORAGE_KEY,
-      `${year}-${month}`
-    );
-  } catch {
-    // Ignore storage failures so the payroll page remains usable.
-  }
 };
 
 const getPayrollJobGroupKey = (jobType: string): string => {
